fix firebase config
